auth-service/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.1.3

# Database
motor>=3.3.0
pymongo>=4.6.0

# HTTP Client
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
